array_merge creates array with merged elements
Added functionality for counting the total
number of elements inside of the multidimensional
array. Also added a function that checks if an
array containing unique values of the multidimensional
array already contains a particular element. Program
calls mergesortRange on array of merged values to be
returned. Most tests are failing